Fix desktop config being lost
A new config file for the menu applets was created after each run in the
desktop setup, because there was no id defined for the applets.

A fixed id was defined for the applets, and the modification that are
made directly to this config files wait for their creating before
running.