Create script to copy back unlinked files
The link to some files are removed when the file is modified by the
application, like the search engine in firefox or the tree style config.

This script copies this files back to the repository, and add back the
symbolic link, in case the link remove behavior is fixed in the
application.